Added folder structure, animations, light mode, mode toggle button. Resolved eslint issues.
