+++ Zplit-Website/.gitignore
@@ -9,6 +9,8 @@
 !.yarn/plugins
 !.yarn/releases
 !.yarn/versions
+*.vscode
+
 
 # testing
 /coverage
