+++ Zplit-Website/app/providers.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import { ThemeProvider } from 'next-themes';
+import { type ReactNode } from 'react';
+
+export function Providers({ children }: { children: ReactNode }) {
+  return (
+    <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+      {children}
+    </ThemeProvider>
+  );
+}
