+++ Zplit-Website/components/home/OpenSourceCTA.tsx
@@ -1,0 +1,91 @@
+'use client';
+
+import { motion } from 'framer-motion';
+
+export default function OpenSourceCTA() {
+  return (
+    <section className="bg-gradient-to-br from-emerald-900 to-teal-900 px-6 py-24 dark:from-emerald-950 dark:to-teal-950">
+      <div className="mx-auto max-w-4xl text-center">
+        <motion.div 
+          initial={{ opacity: 0, scale: 0.8 }}
+          whileInView={{ opacity: 1, scale: 1 }}
+          viewport={{ once: true }}
+          transition={{ duration: 0.6 }}
+          className="mb-6 inline-flex items-center gap-2 rounded-full border border-emerald-400/30 bg-emerald-800/30 px-4 py-2 text-sm font-medium text-emerald-200"
+        >
+          <svg className="h-5 w-5" fill="currentColor" viewBox="0 0 24 24">
+            <path fillRule="evenodd" d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z" clipRule="evenodd" />
+          </svg>
+          100% Open Source
+        </motion.div>
+        
+        <motion.div
+           initial={{ opacity: 0, y: 20 }}
+           whileInView={{ opacity: 1, y: 0 }}
+           viewport={{ once: true }}
+           transition={{ delay: 0.2, duration: 0.6 }}
+        >
+          <h2 className="mb-6 text-4xl font-bold text-white sm:text-5xl">
+            Built in Public, For Everyone
+          </h2>
+          <p className="mb-8 text-lg text-emerald-100">
+            Zplit is completely open source. Contribute, fork, or self-host. No vendor lock-in, no hidden code, no surprises.
+          </p>
+        </motion.div>
+
+        <motion.div 
+          initial={{ opacity: 0, y: 20 }}
+          whileInView={{ opacity: 1, y: 0 }}
+          viewport={{ once: true }}
+          transition={{ delay: 0.4 }}
+          className="flex flex-wrap justify-center gap-4"
+        >
+          <motion.a
+            whileHover={{ scale: 1.05 }}
+            whileTap={{ scale: 0.95 }}
+            href="https://github.com"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="flex items-center gap-2 rounded-full bg-white px-6 py-3 font-semibold text-emerald-900 shadow-lg transition-all hover:shadow-xl"
+          >
+            <svg className="h-5 w-5" fill="currentColor" viewBox="0 0 24 24">
+              <path fillRule="evenodd" d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z" clipRule="evenodd" />
+            </svg>
+            View on GitHub
+          </motion.a>
+          <motion.a
+            whileHover={{ scale: 1.05 }}
+            whileTap={{ scale: 0.95 }}
+            href="https://github.com"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="flex items-center gap-2 rounded-full border-2 border-white px-6 py-3 font-semibold text-white transition-all hover:bg-white hover:text-emerald-900"
+          >
+            ⭐ Star the Project
+          </motion.a>
+        </motion.div>
+        
+        <div className="mt-12 grid gap-6 text-left sm:grid-cols-3">
+          {[
+            { title: "MIT", desc: "Licensed for commercial and personal use" },
+            { title: "Flutter", desc: "Cross-platform with native performance" },
+            { title: "Self-Host", desc: "Deploy your own sync server (optional)" }
+          ].map((item, index) => (
+             <motion.div 
+               key={index}
+               initial={{ opacity: 0, y: 20 }}
+               whileInView={{ opacity: 1, y: 0 }}
+               viewport={{ once: true }}
+               transition={{ delay: 0.6 + (index * 0.1) }}
+               whileHover={{ y: -5 }}
+               className="rounded-2xl border border-emerald-400/20 bg-emerald-800/20 p-6 backdrop-blur"
+             >
+                <div className="mb-2 text-3xl font-bold text-white">{item.title}</div>
+                <div className="text-sm text-emerald-200">{item.desc}</div>
+             </motion.div>
+          ))}
+        </div>
+      </div>
+    </section>
+  );
+}
