--- Zplit-Website/my-app/.vscode/settings.json
@@ -1,1 +1,0 @@
-{}
