+++ Zplit-Website/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-app",
+  "name": "zplit-website",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,7 +9,10 @@
     "lint": "eslint"
   },
   "dependencies": {
+    "framer-motion": "^12.24.12",
+    "lucide-react": "^0.562.0",
     "next": "16.0.10",
+    "next-themes": "^0.4.6",
     "react": "^19",
     "react-dom": "^19"
   },
